Move source and topological api client to ApiClient
