+++ lib/topological_inventory/sync/worker.rb
@@ -2,11 +2,13 @@ require "json"
 require "base64"
 require "manageiq-messaging"
 require "topological_inventory/sync/logging"
+require "topological_inventory/sync/api_client"
 
 module TopologicalInventory
   class Sync
     class Worker
       include Logging
+      include ApiClient
 
       TOPOLOGY_APP_NAME = "/insights/platform/topological-inventory".freeze
 
@@ -76,28 +78,6 @@ module TopologicalInventory
           :service         => queue_name,
           :session_timeout => 60 #seconds
         }
-      end
-
-      def self.sources_api_client(tenant = nil)
-        api_client = SourcesApiClient::ApiClient.new
-        api_client.default_headers.merge!(identity_headers(tenant)) if tenant
-        SourcesApiClient::DefaultApi.new(api_client)
-      end
-
-      def sources_api_client(tenant = nil)
-        self.class.sources_api_client(tenant)
-      end
-
-      def self.identity_headers(tenant)
-        {
-          "x-rh-identity" => Base64.strict_encode64(
-            JSON.dump({"identity" => {"account_number" => tenant}})
-          )
-        }
-      end
-
-      def identity_headers(tenant)
-        self.class.identity_headers(tenant)
       end
     end
   end
